chore(application): Add livewire library and implement basic components
This commit includes the addition of the Livewire library and the implementation of basic components such as the main-content, page-footer, and page-header. Also, updated the composer.json and composer.lock files for the Livewire installation. Besides, it modifies the welcome.blade.php to use the newly implemented Livewire components and updates the Laravel framework version in the composer.lock file. The HRIS logo SVG is also added to the public assets.