chore: Refactor main layout structure and move logic to components
This commit does a major refactor of the main layout structure. The `welcome.blade.php` view was deleted and replaced by the `app.blade.php` file in the `components/layouts` directory. Also, navigation branding and links were moved from `page-header.blade.php` view file to the `PageHeader` livewire component. This way, the logic is directly included within the component.

Moreover, a new `LandingPage` Livewire component was created, with its corresponding view `landing-page.blade.php`. Its purpose is to centralize the logic and display of the landing page. For this reason, the main route in `web.php` file was modified to point to this component instead of the old 'welcome' view.

With this commit, the code keeps a more standard structure, following best practices and improving maintainability.